Rename struct modules to types
